# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Homer-App User interface for Hep protocol",
    "title": "Homer-App",
    "contact": {
      "name": "Aqs",
      "email": "aqsyounas@gmail.com"
    },
    "license": {
      "name": "AGPL",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "host": "localhost:8080",
  "basePath": "/v1",
  "paths": {
...
    "/api/call/report/log": {
      "get": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns logs related to transaction",
        "produces": [
          "application/json"
        ],
        "operationId": "search",
        "responses": {}
      }
    },
    "/api/call/report/qos": {
      "get": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns qos stats",
        "produces": [
          "application/json"
        ],
        "operationId": "search",
        "responses": {}
      }
    },
    "/api/call/transaction": {
      "get": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data related to transaction",
        "produces": [
          "application/json"
        ],
        "operationId": "search",
        "responses": {}
      }
    },
    "/api/search/call/data": {
      "get": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data based upon filtered json",
        "produces": [
          "application/json"
        ],
        "operationId": "search",
        "responses": {}
      }
    },
    "/api/search/call/message": {
      "post": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data based upon filtered json",
        "produces": [
          "application/json"
        ],
        "operationId": "search",
        "responses": {}
      }
    },
...
    "/user": {
      "get": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns the list of Users",
        "produces": [
          "application/json"
        ],
        "tags": [
          "user"
        ],
        "operationId": "ListUsers",
        "responses": {}
      },
      "post": {
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "Create a New user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "user"
        ],
        "operationId": "CreateUser",
        "parameters": [
          {
            "description": "user structure",
            "name": "userstruct",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserStruct"
            }
          }
        ],
        "responses": {}
      }
    },
...
  },
  "definitions": {
...
    "CreateUserStruct": {
      "type": "object",
      "required": [
        "username",
        "partid",
        "email",
        "password",
        "lastname",
        "department",
        "usergroup",
        "guid"
      ],
      "properties": {
        "department": {
          "type": "string",
          "x-go-name": "Department",
          "example": "NOC"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "firstname": {
          "type": "string",
          "x-go-name": "FirstName"
        },
        "guid": {
          "description": "should be a unique value representing user",
          "type": "string",
          "x-go-name": "GUID",
          "example": "e71771a2-1ea0-498f-8d27-391713e10664"
        },
        "lastname": {
          "type": "string",
          "x-go-name": "LastName"
        },
        "partid": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PartId",
          "example": 10
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "usergroup": {
          "type": "string",
          "x-go-name": "UserGroup",
          "example": "admin"
        },
        "username": {
          "type": "string",
          "x-go-name": "UserName"
        }
      },
      "x-go-name": "User",
      "x-go-package": "github.com/sipcapture/homer-app/model"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}